refactor: Moving contact to its own component
